main_app/pyproject.toml
[project]
name = "main_app"
version = "1.0.0"
description = "Веб-приложение на базе фреймворка Litestar с использованием Dependency Injection, SQLAlchemy ORM, RabbitMQ для асинхронной обработки сообщений, Redis для кэширования данных о пользователях, продуктах и заказах, и TaskIQ для планирования и выполнения периодических задач."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "faststream[rabbit]>=0.5.0",
    "litestar>=2.18.0",
    "pika>=1.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.1.1",
    "redis>=5.0.0",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "taskiq>=0.11.0",
    "taskiq-aio-pika>=0.1.0",
    "taskiq-redis>=0.1.0",
    "taskiq-faststream>=0.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "aiosqlite>=0.19.0",
    "polyfactory>=2.0.0",
    "greenlet>=3.0.0",
    "pre-commit>=3.0.0",
    "pylint>=3.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--verbose --color=yes"

[tool.isort]
profile = "black"
known_first_party = ["app"]
skip = []

[tool.black]
line-length = 88
target-version = ['py313']
exclude = '''
/(
  \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
| docker
| config
| temp
| __pycache__
| htmlcov
| tests
| migrations
)/
'''
